==> notebooks/02-Reglas+Clasificación.ipynb <==
# **02-Reglas+Clasificación**

Referencias

* [Pysbd](https://github.com/nipunsadvilkar/pySBD/tree/master)
* [Hyperparameter tuning transformers](https://huggingface.co/docs/transformers/hpo_train)

%cd ..

import json
import numpy as np
import pandas as pd

from sklearn import preprocessing
from sklearn.model_selection import train_test_split

import datasets
from datasets import Dataset, DatasetDict
from transformers import AutoTokenizer, AutoModelForSequenceClassification, TrainingArguments, Trainer
import evaluate
from transformers import pipeline

import pysbd

## Creación del conjunto de datos para entrenar el clasificador

def json_to_df(file_path):

    with open(file_path) as f:
        data = json.load(f)

    entries = []
    for note_id, entry in data["annotated_entries"].items():
        for annotation in entry["section_annotation"]["gold"]:
            d = {
                "text": annotation["segment"],
                "label": annotation["label"],
            }
            entries.append(d)
        
    df = pd.DataFrame(entries)    
    return df

train_path = "data/raw/clinais.train.json"
df_train = json_to_df(train_path)

df_train.to_json("data/rule/train.json")
df_train

le = preprocessing.LabelEncoder()
le.fit(df_train["label"])

df_train["label"] = le.transform(df_train["label"])

df_train

## Búsqueda hiperparámetros transformers

model_checkpoint = "PlanTL-GOB-ES/roberta-base-biomedical-clinical-es"

batch_size = 64

labels = df_train.label.unique().tolist()
train, valid = train_test_split(df_train, test_size=0.2)

dataset = DatasetDict()

dataset["train"] = Dataset.from_pandas(train)
dataset["valid"] = Dataset.from_pandas(valid)
dataset["test"] = Dataset.from_pandas(df_test)

tokenizer = AutoTokenizer.from_pretrained(model_checkpoint, use_fast=True)

def preprocess_function(examples):
    return tokenizer(examples["text"], truncation=True, padding="max_length", max_length=512)

encoded_data = dataset.map(preprocess_function, batched=True)

def model_init():
    return AutoModelForSequenceClassification.from_pretrained(model_checkpoint, num_labels=num_labels, ignore_mismatched_sizes=True)

num_labels = len(set(dataset["train"]["label"]))

metric_acc = evaluate.load("accuracy")

args = TrainingArguments(
    output_dir="data/rule",
    evaluation_strategy = "epoch",
    overwrite_output_dir=True,
    learning_rate=5e-5,
    num_train_epochs=10,
    weight_decay=0.01,
    metric_for_best_model="accuracy",
    load_best_model_at_end=False,
    save_total_limit=10,
    use_mps_device=True,
)

def compute_metrics(eval_pred):
    predictions, labels = eval_pred
    preds = np.argmax(predictions, axis=1)
    return metric_acc.compute(predictions=preds, references=labels)


trainer = Trainer(
    model_init=model_init,
    args=args,
    train_dataset=encoded_data["train"],
    eval_dataset=encoded_data["valid"],
    tokenizer=tokenizer,
    compute_metrics=compute_metrics
)

def hp_space_optuna(trial):
    return {
        "learning_rate": trial.suggest_float("learning_rate", 1e-5, 1e-4),
        "num_train_epochs": trial.suggest_int("num_train_epochs", 1, 3),
        "seed": trial.suggest_int("seed", 1, 40),
        "weight_decay": trial.suggest_float("weight_decay", 0, 0.1),
        "lr_scheduler_type": trial.suggest_categorical("lr_scheduler_type", ["cosine_with_restarts", "linear"]),
    }

best_run = trainer.hyperparameter_search(hp_space=hp_space_optuna, n_trials=10, direction="maximize", backend="optuna") 

best_run

## Creo secciones mediante reglas para el conjunto de test

model_checkpoint = "data/rule/run-0/checkpoint-500"
classifier = pipeline("text-classification", model=model_checkpoint)

test_path = "data/raw/clinais.dev.json"

with open(test_path) as f:
    data = json.load(f)

sections = []
seg = pysbd.Segmenter(language="es", clean=False, char_span=True)

for note_id, entry in data["annotated_entries"].items():
    for sec in seg.segment(entry["note_text"]):
        d = {
            "note_id": note_id,
            "text": sec.sent,
            "label": int(classifier(sec.sent)[0]["label"].replace("LABEL_", "")),
            "start_offset": sec.start,
            "end_offset": sec.end,
        }
        sections.append(d)

df_test = pd.DataFrame(sections)
df_test["label"] = le.inverse_transform(df_test["label"])

df_test

## Transformación de los segmentos y etiquetas para su evaluación

def entry_boundaries(df_test, note_id, boundaries):
    temp_df = df_test[df_test["note_id"]==note_id]
    predictions = boundaries
    for pred in predictions:
        if pred["start_offset"] in temp_df["start_offset"].values:
            pred["boundary"] = temp_df[temp_df["start_offset"]==pred["start_offset"]]["label"].values[0]
        else:
            pred["boundary"] = None
    return predictions   

test_path = "data/raw/clinais.dev.json"

with open(test_path) as f:
    data = json.load(f)

predictions = {}

for note_id, entry in data["annotated_entries"].items():
    predictions[entry["note_id"]] = entry
    predictions[entry["note_id"]]["boundary_annotation"]["prediction"] = entry_boundaries(df_test, note_id, entry["boundary_annotation"]["gold"])
    
with open("data/predictions/predictions_rule_classsification.json", "w") as f:
    json.dump({"annotated_entries": predictions}, f)
