==> notebooks/03-SpanCategorizacer.ipynb <==
# **03-SpanCategorizacer**

Referencias

* [SpanCategorizer](https://spacy.io/api/spancategorizer)
* [Blogpost](https://explosion.ai/blog/spancat)

%cd ..

import json
import spacy
import pandas as pd
import difflib

from sklearn.model_selection import train_test_split
from spacy.tokens import Span, DocBin

nlp = spacy.blank('es')
span_key = "sc"

## Creación del conjunto de datos de entrenamiento

def data_to_spacy_format(file_path):
    with open(file_path) as f:
        data = json.load(f)
    spacy_data = []
    for note_id, entry in data["annotated_entries"].items():
        doc = nlp(entry["note_text"])
        spans = []
        start_span = 1
        for annotation in entry["section_annotation"]["gold"]:
            seg = nlp(annotation["segment"])
            spans.append(Span(doc, start_span, start_span+len(seg), annotation["label"]))
        doc.spans[span_key] = spans
        spacy_data.append(doc)
    return spacy_data

train_path = "data/raw/clinais.train.json"
train_data = data_to_spacy_format(train_path)

train, valid = train_test_split(train_data, test_size=0.2)

DocBin(docs=train).to_disk("data/spacy/train.spacy")
DocBin(docs=valid).to_disk("data/spacy/valid.spacy")

## Entrenamiento del modelo

! python -m spacy init fill-config data/spacy/base_config.cfg data/spacy/base_config.cfg

! python -m spacy train data/spacy/base_config.cfg --output data/spacy/output --paths.train data/spacy/train.spacy --paths.dev data/spacy/valid.spacy --gpu-id 0

## Predicción sobre nuevos datos

nlp = spacy.load("data/spacy/output/model-best")
nlp_es = spacy.load("es_core_news_sm")

test_path = "data/raw/clinais.dev.json"

with open(test_path) as f:
    data = json.load(f)

annotations = []

for note_id, entry in data["annotated_entries"].items():
    text = entry["note_text"]

    doc = nlp(text)
    spans = doc.spans[span_key]
    spans_dict = {span.text: span.label_ for span in spans}

    start_offset = 0
    
    for sent in nlp_es(text).sents:
        if sent in spans_dict:
            label = spans_dict[sent.text]
        else:
            most_similar = difflib.get_close_matches(sent.text, spans_dict.keys(), n=1)[0]
            label = spans_dict[most_similar]

        d = {
            "note_id": note_id,
            "segment": sent.text,
            "label": label,
            "start_offset": start_offset
        }
        annotations.append(d)
        start_offset += len(sent.text)

df = pd.DataFrame(annotations)

df  

def entry_boundaries(df_test, note_id, boundaries):
    temp_df = df_test[df_test["note_id"]==note_id]
    predictions = boundaries
    for pred in predictions:
        if pred["start_offset"] in temp_df["start_offset"].values:
            pred["boundary"] = temp_df[temp_df["start_offset"]==pred["start_offset"]]["label"].values[0]
        else:
            pred["boundary"] = None
    return predictions   

predictions = {}

for note_id, entry in data["annotated_entries"].items():
    predictions[entry["note_id"]] = entry
    predictions[entry["note_id"]]["boundary_annotation"]["prediction"] = entry_boundaries(df, note_id, entry["boundary_annotation"]["gold"])
    
with open("data/predictions/predictions_spacy_spancat.json", "w") as f:
    json.dump({"annotated_entries": predictions}, f)
