==> notebooks/04.2-NER-Bert.ipynb <==
# **04.2-NER-Bert**

References:
* [Token Classification HuggingFace](https://huggingface.co/learn/nlp-course/chapter7/2)

%cd ..

import json
import re
import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split

import datasets
from datasets import Dataset, DatasetDict
from transformers import AutoTokenizer, AutoModelForTokenClassification, DataCollatorForTokenClassification, TrainingArguments, Trainer
from transformers import pipeline

import evaluate

def json_to_df(file_path, label2id):

    with open(file_path) as f:
        data = json.load(f)

    entries = []
    for note_id, entry in data["annotated_entries"].items():
        text = entry["note_text"]
        tokens = re.findall(r"\w+|[^\w\s]", text, re.UNICODE)
        labels = np.zeros(len(tokens))
        position = 0
        for annotation in entry["section_annotation"]["gold"]:
            segment = annotation["segment"]
            segment_tokens = re.findall(r"\w+|[^\w\s]", segment, re.UNICODE)
            labels[position] =  label2id[annotation["label"]]
            position += len(segment_tokens)
            
        d = {
            "tokens": tokens,
            "ner_tags": labels.astype(int),
        }    
        entries.append(d)
        
    df = pd.DataFrame(entries)    
    return df

label2id = {
    "O": 0,
    "EXPLORATION": 1,
    "TREATMENT": 2,
    "PRESENT_ILLNESS": 3,
    "EVOLUTION": 4,
    "PAST_MEDICAL_HISTORY": 5,
    "DERIVED_FROM/TO": 6,
    "FAMILY_HISTORY": 7,
}

id2label = {v: k for k, v in label2id.items()}

train_path = "data/raw/clinais.train.json"
df_train = json_to_df(train_path, label2id)

df_train

train, valid = train_test_split(df_train, test_size=0.2)

dataset = DatasetDict()

dataset["train"] = Dataset.from_pandas(train, preserve_index=False)
dataset["valid"] = Dataset.from_pandas(valid, preserve_index=False)

def align_labels_with_tokens(labels, word_ids):
    new_labels = []
    current_word = None
    for word_id in word_ids:
        if word_id != current_word:
            # Start of a new word!
            current_word = word_id
            label = -100 if word_id is None else labels[word_id]
            new_labels.append(label)
        elif word_id is None:
            # Special token
            new_labels.append(-100)
        else:
            # Same word as previous token
            label = labels[word_id]
            # If the label is B-XXX we change it to I-XXX
            new_labels.append(label)

    return new_labels

def tokenize_and_align_labels(examples):
    tokenized_inputs = tokenizer(
        examples["tokens"], truncation=True, is_split_into_words=True
    )
    all_labels = examples["ner_tags"]
    new_labels = []
    for i, labels in enumerate(all_labels):
        word_ids = tokenized_inputs.word_ids(i)
        new_labels.append(align_labels_with_tokens(labels, word_ids))

    tokenized_inputs["labels"] = new_labels
    return tokenized_inputs

model_checkpoint = "PlanTL-GOB-ES/roberta-base-biomedical-clinical-es"

tokenizer = AutoTokenizer.from_pretrained(model_checkpoint)
tokenized_dataset = dataset.map(tokenize_and_align_labels, batched=True, remove_columns=dataset["train"].column_names)

data_collator = DataCollatorForTokenClassification(tokenizer=tokenizer)

def model_init():
    return AutoModelForTokenClassification.from_pretrained(model_checkpoint, id2label=id2label, label2id=label2id)

args = TrainingArguments(
    output_dir="data/NER/model",
    evaluation_strategy = "epoch",
    overwrite_output_dir=True,
    learning_rate=5e-5,
    num_train_epochs=10,
    weight_decay=0.01,
    metric_for_best_model="Weighted B2",
    load_best_model_at_end=False,
    save_total_limit=10,
    use_mps_device=True,
)

metric = evaluate.load("seqeval")

def compute_metrics(eval_preds):
    logits, labels = eval_preds
    predictions = np.argmax(logits, axis=-1)

    # Remove ignored index (special tokens) and convert to labels
    true_labels = [[id2label[l] for l in label if l != -100] for label in labels]
    true_predictions = [
        [id2label[p] for (p, l) in zip(prediction, label) if l != -100]
        for prediction, label in zip(predictions, labels)
    ]
    all_metrics = metric.compute(predictions=true_predictions, references=true_labels)
    return {
        "precision": all_metrics["overall_precision"],
        "recall": all_metrics["overall_recall"],
        "f1": all_metrics["overall_f1"],
        "accuracy": all_metrics["overall_accuracy"],
    }

from src.evaluation.evaluate import evaluate_note
from src.evaluation.dataset_model import BoundaryAnnotation
from numpy import average

id2label[0]=None

def compute_metrics(eval_preds):
    logits, labels = eval_preds
    predictions = np.argmax(logits, axis=-1)

    # Remove ignored index (special tokens) and convert to labels
    true_labels = [[l for l in label if l != -100] for label in labels]
    true_predictions = [
        [p for (p, l) in zip(prediction, label) if l != -100]
        for prediction, label in zip(predictions, labels)
    ]
    number_sections_per_file = []
    scores_per_file = []

    # Compute the scores
    for label, prediction in zip(true_labels, true_predictions):
        number_sections, metrics = evaluate_note(
            reference_boundaries=[BoundaryAnnotation(span=" ", start_offset=i, end_offset=i+1, boundary=id2label[l]) for i, l in enumerate(label)],
            prediction_boundaries=[BoundaryAnnotation(span=" ", start_offset=i, end_offset=i+1, boundary=id2label[l]) for i, l in enumerate(prediction)]
            #reference_boundaries=[{"span": " ", "start_offset": i, "end_offset": i+1, "boundary": l} for i, l in enumerate(label)],
            #prediction_boundaries=[{"span": " ", "start_offset": i, "end_offset": i+1, "boundary": l} for i, l in enumerate(prediction)]
        )
        number_sections_per_file.append(number_sections)
        scores_per_file.append(metrics["metrics"]["B2"])

    # Compute the final scores on the whole dataset
    return {"Weighted B2": average(scores_per_file, weights=number_sections_per_file)}

trainer = Trainer(
    model_init=model_init,
    args=args,
    train_dataset=tokenized_dataset["train"],
    eval_dataset=tokenized_dataset["valid"],
    data_collator=data_collator,
    compute_metrics=compute_metrics,
    tokenizer=tokenizer,
)

def hp_space_optuna(trial):
    return {
        "learning_rate": trial.suggest_float("learning_rate", 1e-5, 1e-4),
        "num_train_epochs": trial.suggest_int("num_train_epochs", 1, 3),
        "seed": trial.suggest_int("seed", 1, 40),
        "weight_decay": trial.suggest_float("weight_decay", 0, 0.1),
        "lr_scheduler_type": trial.suggest_categorical("lr_scheduler_type", ["cosine_with_restarts", "linear"]),
    }

best_run = trainer.hyperparameter_search(hp_space=hp_space_optuna, n_trials=10, direction="maximize", backend="optuna") 

best_run

model_checkpoint = "data/NER/model/checkpoint-500"
token_classifier = pipeline("token-classification", model=model_checkpoint, aggregation_strategy="simple")

test_path = "data/raw/clinais.dev.json"

with open(test_path) as f:
    data = json.load(f)

annotations = []

for note_id, entry in data["annotated_entries"].items():
    text = entry["note_text"]
    ents = token_classifier(text)

    for ent in ents:
        d = {
            "note_id": note_id,
            "label": ent["entity_group"],
            "start_offset": ent["start"]
        }
        annotations.append(d)

df = pd.DataFrame(annotations)

df

def entry_boundaries(df_test, note_id, boundaries):
    temp_df = df_test[df_test["note_id"]==note_id]
    predictions = boundaries
    for pred in predictions:
        if pred["start_offset"] in temp_df["start_offset"].values:
            pred["boundary"] = temp_df[temp_df["start_offset"]==pred["start_offset"]]["label"].values[0]
        else:
            pred["boundary"] = None
    return predictions   

predictions = {}

for note_id, entry in data["annotated_entries"].items():
    predictions[entry["note_id"]] = entry
    predictions[entry["note_id"]]["boundary_annotation"]["prediction"] = entry_boundaries(df, note_id, entry["boundary_annotation"]["gold"])
    
with open("data/predictions/predictions_ner_bert.json", "w") as f:
    json.dump({"annotated_entries": predictions}, f) 

No puedo aplicar la métrica B2 porque en compute_metrics solo tengo la lista de etiquetas y predicciones de cada token, y necesitaría saber también la posición inicial de cada token para calcular cuánto se ha ido. La he implementado poniendo que cada token es de longitud 1.

Utilizando como métrica el accuracy 10 epochs "Weighted B2": 0.5340219745589161
Utilizando como métrica b2 10 epochs "Weighted B2": 0.5877273720523583,
