==> notebooks/05-Evaluación.ipynb <==
# **05-Evaluación**

%cd ..

import json

from src.evaluation import evaluate

evaluate.score_predictions(prediction_file="data/predictions/predictions_rule_classsification.json", reference_file="data/raw/clinais.dev.json", output_result_file="data/rule/results_rule_classification.json")

evaluate.score_predictions(prediction_file="data/predictions/predictions_spacy_spancat.json", reference_file="data/raw/clinais.dev.json", output_result_file="data/spacy/results_spacy_spancat.json")

evaluate.score_predictions(prediction_file="data/predictions/predictions_spacy_ner.json", reference_file="data/raw/clinais.dev.json", output_result_file="data/ner/results_spacy_ner.json")

evaluate.score_predictions(prediction_file="data/predictions/predictions_ner_bert.json", reference_file="data/raw/clinais.dev.json", output_result_file="data/results/results_ner_bert.json")
